# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Loup — household services marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/admin/webhook-events/{id}/replay:
    post:
      operationId: replayAdminWebhookEvent
      tags: [platform]
      summary: Re-enqueue a webhook event for delivery
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event re-enqueued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookReplayResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /v1/admin/webhook-events/purge:
    post:
      operationId: purgeAdminWebhookEvents
      tags: [platform]
      summary: PDPL data-minimization retention purge — deletes resolved (delivered/failed) webhook events older than the retention window (P1-11)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookPurgeInput"
      responses:
        "200":
          description: Purge result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookPurgeResult"
  /v1/admin/webhook-endpoints:
    get:
      operationId: listAdminWebhookEndpoints
      tags: [platform]
      summary: List webhook delivery endpoints across institutions
      responses:
        "200":
          description: Webhook endpoints
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEndpointsResult"
    post:
      operationId: createAdminWebhookEndpoint
      tags: [platform]
      summary: Register a webhook delivery endpoint for an institution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookEndpointInput"
      responses:
        "201":
          description: Endpoint created (signing secret returned once)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEndpointCreated"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /v1/admin/webhook-endpoints/{id}:
    delete:
      operationId: deactivateAdminWebhookEndpoint
      tags: [platform]
      summary: Deactivate a webhook delivery endpoint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Endpoint deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEndpointDeactivateResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /v1/widget/token:
    post:
      operationId: getWidgetToken
      tags: [platform]
      summary: Server-to-server exchange — issues a short-lived employee-scoped token for the embeddable widget (P1-7)
      parameters:
        - name: X-Loup-Widget-Secret
          in: header
          required: true
          schema:
            type: string
          description: Per-institution widget secret, issued out of band — never sent from the browser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WidgetTokenInput"
      responses:
        "200":
          description: Short-lived employee token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetWidgetTokenResult"
        "401":
          description: Invalid widget secret
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "404":
          description: Unknown institution or employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "503":
          description: Widget not provisioned for this institution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
components:
  schemas:
...
    WebhookPurgeInput:
      type: object
      properties:
        olderThanDays: { type: integer, minimum: 1, description: "Defaults to a 90-day retention window server-side when omitted" }
    WebhookPurgeResult:
      type: object
      required: [purged, retentionDays]
      properties:
        purged: { type: integer }
        retentionDays: { type: integer }
...
    ApiMessage:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    WebhookReplayResult:
      type: object
      required: [ok, id, status, nextAttemptAt, maxAttempts, backoffSeconds, payload]
      properties:
        ok: { type: boolean }
        id: { type: integer }
        status: { type: string }
        nextAttemptAt: { type: string }
        maxAttempts: { type: integer }
        backoffSeconds: { type: number }
        payload: { type: object, additionalProperties: true }
    WebhookEndpoint:
      type: object
      required: [id, institutionId, institutionName, url, active, createdAt]
      properties:
        id: { type: integer }
        institutionId: { type: integer }
        institutionName: { type: string }
        url: { type: string }
        active: { type: boolean }
        createdAt: { type: string }
    WebhookEndpointsResult:
      type: object
      required: [endpoints]
      properties:
        endpoints:
          type: array
          items:
            $ref: "#/components/schemas/WebhookEndpoint"
    WebhookEndpointInput:
      type: object
      required: [institutionId, url]
      properties:
        institutionId: { type: integer }
        url: { type: string, minLength: 1 }
    WebhookEndpointCreated:
      type: object
      required: [id, institutionId, url, active, secret]
      properties:
        id: { type: integer }
        institutionId: { type: integer }
        url: { type: string }
        active: { type: boolean }
        secret:
          type: string
          description: Signing secret for HMAC verification — shown only once, on creation.
    WebhookEndpointDeactivateResult:
      type: object
      required: [ok, id]
      properties:
        ok: { type: boolean }
        id: { type: integer }
    WidgetTokenInput:
      type: object
      required: [institutionSlug, externalEmployeeId]
      properties:
        institutionSlug: { type: string }
        externalEmployeeId: { type: string, description: "The employer's own employee identifier, as synced via the roster (P1-6)" }
    GetWidgetTokenResult:
      type: object
      required: [token, expiresInSeconds, employeeName]
      properties:
        token: { type: string }
        expiresInSeconds: { type: integer }
        employeeName: { type: string }
...