# lib/api-zod/src/generated/types/webhookEndpoint.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Loup — household services marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface WebhookEndpoint {
  id: number;
  institutionId: number;
  institutionName: string;
  url: string;
  active: boolean;
  createdAt: string;
}