# lib/api-zod/src/generated/types/webhookEndpointCreated.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Loup — household services marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface WebhookEndpointCreated {
  id: number;
  institutionId: number;
  url: string;
  active: boolean;
  /** Signing secret for HMAC verification — shown only once, on creation. */
  secret: string;
}