# lib/api-zod/src/generated/types/webhookEndpointDeactivateResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Loup — household services marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface WebhookEndpointDeactivateResult {
  ok: boolean;
  id: number;
}