# lib/api-zod/src/generated/types/webhookEndpointInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Loup — household services marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface WebhookEndpointInput {
  institutionId: number;
  /** @minLength 1 */
  url: string;
}