# lib/api-zod/src/generated/types/webhookEndpointsResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Loup — household services marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { WebhookEndpoint } from './webhookEndpoint';

export interface WebhookEndpointsResult {
  endpoints: WebhookEndpoint[];
}